requirements.txt
# Dutch-o-matic - Requirements
# Python 3.8+

# Core dependencies
python-escpos>=3.0
pyserial>=3.5
pillow>=9.0.0
flask>=2.0.0
